feat(visits): add services for fetching past and upcoming visits, update controller and routes
