+++ src/controller/visits.client.controller.js
@@ -1,6 +1,6 @@
 import mongoose from "mongoose"
 import { Visit } from "../model/visit.model.js"
-import { createVisitService, getVisits, updateVisitService } from "../services/visit.services.js"
+import { createVisitService, getPastVisitsService, getUpcomingVisitsService, getVisits, updateVisitService } from "../services/visit.services.js"
 
 export const createVisit = async (req, res, next) => {
 
@@ -81,26 +81,7 @@ export const getPastVisits = async (req, res, next) => {
     const client = req.user._id
 
     try {
-        const visits = await Visit.find({
-            client,
-            date: { $lt: new Date() }
-        })
-            .sort({ date: -1 }) // most recent first
-            .skip((page - 1) * limit)
-            .limit(Number(limit));
-
-        const total = await Visit.countDocuments({
-            client,
-            date: { $lt: new Date() }
-        });
-
-        return res.status(200).json({
-            status: true,
-            message: "Past visits fetched successfully",
-            data: visits,
-            totalPages: Math.ceil(total / limit),
-            currentPage: Number(page)
-        });
+        await getPastVisitsService(page, limit, client, res)
     }
 
     catch (error) {
@@ -113,26 +94,7 @@ export const getUpcomingVisits = async (req, res, next) => {
     const client = req.user._id
 
     try {
-        const visits = await Visit.find({
-            client,
-            date: { $gte: new Date() }
-        })
-            .sort({ date: 1 })
-            .skip((page - 1) * limit)
-            .limit(Number(limit));
-
-        const total = await Visit.countDocuments({
-            client,
-            date: { $gte: new Date() }
-        });
-
-        return res.status(200).json({
-            status: true,
-            message: "Upcoming visits fetched successfully",
-            data: visits,
-            totalPages: Math.ceil(total / limit),
-            currentPage: Number(page)
-        });
+        await getUpcomingVisitsService(page, limit, client, res)
     }
 
     catch (error) {
