+++ src/routes/visits.admin.route.js
@@ -8,12 +8,12 @@ const router = Router();
 router.post("/create-visit", verifyJWT, isAdmin, createVisit);
 router.get('/get-all-visits-count', verifyJWT, isAdmin, getAllVisitsCount);
 router.get('/get-pending-visits-count', verifyJWT, isAdmin, getPendingVisitsCount)
-router.get('/get-confirmed-visits', verifyJWT, isAdmin, getConfirmedVisits);
-router.get('/get-pending-visits', verifyJWT, isAdmin, getPendingVisits);
-router.get('/get-completed-visits', verifyJWT, isAdmin, getCompletedVisits);
-router.get('/get-cancelled-visits', verifyJWT, isAdmin, getCancelledVisits)
-router.get('/get-past-visits', verifyJWT, isAdmin, getPastVisits)
-router.get('/get-upcoming-visits', verifyJWT, isAdmin, getUpcomingVisits)
+router.get('/get-confirmed-visits/:client', verifyJWT, isAdmin, getConfirmedVisits);
+router.get('/get-pending-visits/:client', verifyJWT, isAdmin, getPendingVisits);
+router.get('/get-completed-visits/:client', verifyJWT, isAdmin, getCompletedVisits);
+router.get('/get-cancelled-visits/:client', verifyJWT, isAdmin, getCancelledVisits)
+router.get('/get-past-visits/:client', verifyJWT, isAdmin, getPastVisits)
+router.get('/get-upcoming-visits/:client', verifyJWT, isAdmin, getUpcomingVisits)
 router.patch('/update-visit/:id', verifyJWT, isAdmin, updateVisit);
 router.patch('/update-visit-staff/:id', verifyJWT, isAdmin, updateVisitStaff);
